transfer WMFPcalc func declaration to pkg/kys
