+++ cmd/gokys/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"flag"
 	"fmt"
-	"go/ast"
+	"github.com/bragov4ik/go-kys/pkg/kys"
 	"go/parser"
 	"go/token"
 	"log"
@@ -43,7 +43,7 @@ func main() {
 		if err != nil {
 			log.Fatal(err)
 		}
-		WMFPcalc(node, scores)
+		kys.WMFPcalc(node, scores)
 	}
 
 	fmt.Println(scores)
@@ -70,33 +70,4 @@ func WalkMatch(root, pattern string) ([]string, error) {
 		return nil, err
 	}
 	return matches, nil
-}
-
-func WMFPcalc(file *ast.File, score map[string]int) {
-	ast.Inspect(file, func(n ast.Node) bool {
-		// Find function declarations
-		_, ok := n.(*ast.FuncDecl)
-		if ok {
-			score["funcDecl"]++
-			return true
-		}
-		// Find return statements
-		_, ok = n.(*ast.ReturnStmt)
-		if ok {
-			score["returnStmt"]++
-			return true
-		}
-		// Find function calls
-		_, ok = n.(*ast.CallExpr)
-		if ok {
-			score["callExpr"]++
-			return true
-		}
-		// Find assignment statements
-		_, ok = n.(*ast.AssignStmt)
-		if ok {
-			score["assignStmt"]++
-		}
-		return true
-	})
 }
